Restructure folders for new landing page
