+++ website/src/app/[lang]/[region]/(website)/v2/home/(sections)/sdgs.tsx
